Added missing opcodes as comments to opcodes.h
Note that the list is not perfect. Some opcodes (such as ALOAD) are
not on the list, but are supported since they get translated by
the .class file translation tool.
Some others are in the list (ALOAD0 to ALOAD3), but are never used
since they get translated to their ILOAD counterparts.