Added several testcases to check out some strange behaviour on IO ports.
In each test all for bits of an IO port (either port A or port C) were
turned on.
To do this we first call AVR.port[A|C].setOuput(x), and then
AVR.port[A|C].setBit(x), with x=0 to 7.
All combinations of the following 4 options were tried:
 - Printing 2 lines before turning on the port, or not
 - Using the AVR.portA directly, or storing it in a local variable
 - Using a for loop to set all 8 bits, or unrolling the loop
 - Using port A or port C