Work in progress.
We can now change code, but it turns out the bytecode is running
in RAM rather than EEPROM, but changing the config option to use
EEPROM instead gives a verification error when uploading with
avrdude. Using Flash is also tricky since we need to erase/write
whole pages instead of single bytes.