Removed monitor from Arduino bootloader.
Not sure how it's supposed to be used, but avrdude works fine
without it, and dropping this feature leaves ample space for our
own stuff in the NRWW section.
Compile with "make clean" "make mega2560". (the default mega128
doesn't seem to work, and it won't build without a "make clean"
either)