Added make target for smaller bootloader and our own Hello World in the NRWW.
Use "make upload_bootloader" to upload the Arduino bootloader
without the "monitor", and add our own code to the NRWW section.
The bootloader is in the /arduino/arduino-stk500v2-bootloader
directory and has to be compiled first, and the Arduino has to
be connected to the USBtinyISP programmer.
After that "make avrdude" will use the normal Arduino bootloader
to load the VM, which will call the hello() function in the NRWW
to display a Hello World message on startup.