Extremely dirty hack to get the code update working.
To get around the problem of how to align the bytecode in flash,
I defined a second array nvmfile_flash, which is put in a separate
section called .javabytecode. This section is hardcoded at address
0x4000.
I cannot put nvmfile there itself, since this is a variable in SRAM
that needs to be initialised and it won't be if I place it in
another section. (the VM currently executes bytecode from SRAM,
whether or not this is what we want is another question)

So instead I place nvmfile_flash at a fixed address in flash, and
initialise nvmfile myself by using pgm_read_byte to copy the data
from flash to SRAM. This is done in nvmfile_init().

I'm sure there's a better way to do this, but for now at least we
can demo some sort of code update.