Thin Python wrapper around testrtt.c. Can receive messages from nodes now.
PyZwave is a very minimal wrapper around testrtt.c. It needs to be initialised
using
pyzwave.init(<IP address of the HSC-30>), then we can send using
pyzwave.senddata(dest, data), or receive data using
pyzwave.receive(<timeout in ms>)
Work in progress: for the moment the receive function just gives you the data,
not even the ID of the sending node yet.