Can send and receive from Java.
NvmComm3.receive takes the number of milliseconds to wait for a message
as a parameter, and returns either NULL or an array containing the
received data. The TestCodeUpdate.java in this commit waits for a
message (type NVC3_CMD_APPMSG) and does a Rot13 on the contents, which
it then sends back.
This works with the Python library and a test in reprogram.py. However,
it also turns out there's a bug in my pyzwave library which causes it
to fail when it wants to send a second message. This means the actual
code update doesn't work anymore until I fix this issue.