Renamed NRWW section to .reprogram_flash_page
It was called .bootloader before since the function was copied
from the example in the AVR's boot.h header file. But this name is
a bit confusing since it has nothing to do with the actual Arduino
bootloader that sits besides it in the NRWW section. Instead it's
just a single function that gets called by the VM if it receives
a new byte code image.