Added features to configuration code. Can be accessed through two new nvmcomm message types.
Each sensor or native software class can be a feature. Every feature is a bit in EEPROM. In wkpf_native_wuclasses_init each WuClass is only registered/created when the corresponding bit is set. This means the pin to which sensors have to be connected is still hardcoded, and consequently that we can't have multiple sensors of the same type on the same node, but I think for now this will be enough configurability.
TODO:
 * solve the issue with WuObjects not being created for hardware (why was this changed in the first place?)
 * add some python code to handle the new nvmcomm message types