Fixed writing to eeprom using the boot loader
This feature was simply commented out in the original boot loader. Probably because it's using the EEWE and EEMWE constants, which stopped working after they were renamed to EEPE/EEPWE.
Renaming those in the boot loader would probably have worked, but since I'm not too sure about the best way to write to eeprom (other code I found online mentions disabling interrupts, which the original boot loader doesn't seem to do), I decided to use the functions from <avr/eeprom.h> instead.

It seems to work perfectly now, but in a first test something went wrong and after writing the first 0x60 bytes correctly, the next one ended up at address 0x90 instead of 0x60. Not sure if that was my own mistake, but to be sure I enabled verification in the makefile when uploading with avrdude. This makes uploading code take slightly longer, but given how extremely buggy the toolchain is this seems a reasonable price to pay for some extra checks.