Fixed problem with missing eeprom_update_ functions when using older version of AVR libs.
Using nanokong's eeprom.h now instead of avr/eeprom.h directly. This should also make it work on the unix target as well and make it easier to port to other platforms by implementing the right functions in eeprom.h.
If the eeprom_update_ functions are missing, they get #defined to the eeprom_write_ equivalent.